+++ lib/mollie_ex/resources/organizations/current.ex
@@ -1,0 +1,33 @@
+defmodule MollieEx.Resources.Organizations.Current do
+  @moduledoc false
+
+  alias MollieEx.Client
+  alias MollieEx.Error
+  alias MollieEx.HTTP.Request
+  alias MollieEx.Resources.Options
+  alias MollieEx.Resources.RequestBuilder
+
+  @allowed_options [
+    :pool_timeout,
+    :receive_timeout,
+    :request_timeout
+  ]
+
+  @spec build(Client.t(), keyword()) ::
+          {:ok, Request.t(), keyword()} | {:error, Error.t()}
+  def build(%Client{} = client, opts) when is_list(opts) do
+    with :ok <- Options.validate_options(opts, @allowed_options),
+         :ok <- Options.reject_api_key_client(client) do
+      RequestBuilder.build(opts,
+        method: :get,
+        path: "/organizations/me",
+        path_template: "/organizations/me",
+        query: [],
+        idempotency_policy: :unsupported,
+        operation: :organizations_current
+      )
+    end
+  end
+
+  def build(%Client{}, _opts), do: Options.configuration_error(:invalid_options)
+end